Crac results variant Management
