+++ data/crac/crac-result-extensions/src/test/java/com.farao_community.farao.data.crac_result_extensions/CracResultTest.java
@@ -20,12 +20,6 @@ public class CracResultTest {
     private double epsilon = 0.0;
 
     @Test
-    public void getName() {
-        CracResult cracResult = new CracResult();
-        assertEquals("CracResult", cracResult.getName());
-    }
-
-    @Test
     public void securedCracResult() {
         double cost = -5;
         CracResult securedCracResult = new CracResult(cost);
