+++ data/crac/crac-result-extensions/src/test/java/com.farao_community.farao.data.crac_result_extensions/RangeActionResultsExtensionTest.java
@@ -23,7 +23,7 @@ import static org.junit.Assert.assertNull;
 /**
  * @author Baptiste Seguinot {@literal <baptiste.seguinot at rte-france.com>}
  */
-public class RangeActionResultsExtensionsTest {
+public class RangeActionResultsExtensionTest {
 
     @Test
     public void testVariantManagementOk() {
